Experiments with sine algorithms
