+++ ExtendedCModule_VS22.sln
@@ -19,6 +19,8 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ecm.math", "Libs\ecm.math.v
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ecm.opengl", "Libs\ecm.opengl.vcxproj", "{259699BF-2EDF-430E-85DB-0B429255BFC4}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Math", "TestApps\Math\Math.vcxproj", "{DA99B498-ECF2-40B1-A6BB-8CD3660F61F0}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|x64 = Debug|x64
@@ -75,6 +77,14 @@ Global
 		{259699BF-2EDF-430E-85DB-0B429255BFC4}.Release|x64.Build.0 = Release|x64
 		{259699BF-2EDF-430E-85DB-0B429255BFC4}.Release|x86.ActiveCfg = Release|Win32
 		{259699BF-2EDF-430E-85DB-0B429255BFC4}.Release|x86.Build.0 = Release|Win32
+		{DA99B498-ECF2-40B1-A6BB-8CD3660F61F0}.Debug|x64.ActiveCfg = Debug|x64
+		{DA99B498-ECF2-40B1-A6BB-8CD3660F61F0}.Debug|x64.Build.0 = Debug|x64
+		{DA99B498-ECF2-40B1-A6BB-8CD3660F61F0}.Debug|x86.ActiveCfg = Debug|Win32
+		{DA99B498-ECF2-40B1-A6BB-8CD3660F61F0}.Debug|x86.Build.0 = Debug|Win32
+		{DA99B498-ECF2-40B1-A6BB-8CD3660F61F0}.Release|x64.ActiveCfg = Release|x64
+		{DA99B498-ECF2-40B1-A6BB-8CD3660F61F0}.Release|x64.Build.0 = Release|x64
+		{DA99B498-ECF2-40B1-A6BB-8CD3660F61F0}.Release|x86.ActiveCfg = Release|Win32
+		{DA99B498-ECF2-40B1-A6BB-8CD3660F61F0}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -86,6 +96,7 @@ Global
 		{49A18B56-B187-4F1E-AE63-C2D2A9A51BBA} = {25980BF8-0D7A-4EF6-B935-CEEE06CC2240}
 		{E86FB754-590E-4D28-B5BB-BE344AC043B4} = {25980BF8-0D7A-4EF6-B935-CEEE06CC2240}
 		{259699BF-2EDF-430E-85DB-0B429255BFC4} = {25980BF8-0D7A-4EF6-B935-CEEE06CC2240}
+		{DA99B498-ECF2-40B1-A6BB-8CD3660F61F0} = {32F7E4A6-FC11-4D79-B824-B54FD9AB31A9}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {4B6EB90A-279D-4E4B-A044-2093628F28FE}
