+++ Libs/ecm/ecm_configs.h
@@ -9,11 +9,13 @@
 
 #if ECM_ALL_DYNAMIC_BUILD
 #define ECM_DYNAMIC_BUILD 1
+#define ECM_MATH_DYNAMIC_BUILD 1
 #define ECM_ALG_DYNAMIC_BUILD 1
 #define ECM_WIN_DYNAMIC_BUILD 1
 #define ECM_GL_DYNAMIC_BUILD 1
 #else
 #define ECM_DYNAMIC_BUILD 0
+#define ECM_MATH_DYNAMIC_BUILD 0
 #define ECM_ALG_DYNAMIC_BUILD 0
 #define ECM_WIN_DYNAMIC_BUILD 0
 #define ECM_GL_DYNAMIC_BUILD 0
